+++ docs/ReleaseNotes.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [3.6.2] - 2024.07.26 unpublished
+## [3.7.1] - 2024.07.26 unpublished
 
 - Minor fixes and code formatting improvements.
 - Added information message.
 - Improve cleaning function.
-- Fix passing of maxtokens value
+- Fix passing of maxtokens value.
+- Added new function `Test-NeedForMoreInfo` to assess if additional information is required for processing.
 
 ## [3.6.1] - 2024.07.25